requirements.txt
python-slugify==1.2.6
hdx-python-api==3.4.9
-r docker-requirements.txt